+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 # Changelog
 
+## v1.1.0
+- Adds better check on `method` option when creating a `Route`
+- Adds `server` as the 4th argument passed to middleware and route handler
+- Adds `Server.autoRestart()`
+- Adds `Server.setPort(Number)`
+- Applies the option `formatJson` to all routes using a middleware
+- Changes `formatJson` to `true` in `Server` options
+
 ## v1.0.0
 - First public release